anpr_platform/requirements.txt
fastapi
uvicorn
paho-mqtt
sqlalchemy
websockets
jinja2
python-multipart
